feat: Integrate metadata abstractions with HiveServer2 GetObjects
Replace TableInfo with TableMetadata + ColumnMetadataRecord in GetObjects:
- Replace Dictionary<..., TableInfo> with Dictionary<..., TableMetadata>
- Use MetadataFieldPopulator to create ColumnMetadataRecords
- Replace GetColumnSchema with MetadataSchemaBuilder.BuildColumnsStructArray
- Preserve exact same output values and behavior

Key changes:
- GetObjects now uses shared ColumnMetadataRecord data model
- Type mapping logic unified through ColumnTypeMapper
- Schema building delegated to MetadataSchemaBuilder
- Zero breaking changes - byte-for-byte identical output

This enables code reuse with StatementExecution API in Phase 3.

+++ csharp/src/Drivers/Apache/Hive2/Metadata/MetadataSchemaBuilder.cs
@@ -16,6 +16,7 @@
 */
 
 using System.Collections.Generic;
+using Apache.Arrow.Adbc.Extensions;
 using Apache.Arrow.Types;
 
 namespace Apache.Arrow.Adbc.Drivers.Apache.Hive2.Metadata
@@ -548,7 +549,7 @@ namespace Apache.Arrow.Adbc.Drivers.Apache.Hive2.Metadata
         private static void AppendOrNull(Int8Array.Builder builder, byte? value)
         {
             if (value.HasValue)
-                builder.Append(value.Value);
+                builder.Append((sbyte)value.Value);
             else
                 builder.AppendNull();
         }
